feat(portal): merge API Keys + Docs into single Agentic Access page nested under Events
Collapse the two Agentic Access subroutes into /portal/agentic-access and
nest it under Events Calendar in the sidebar, so users mint a key and read
the API reference in one scroll. Surface the agent brief copy button in
the post-creation modal too.
