--- portal/src/app/portal/docs/page.tsx
@@ -1,459 +1,0 @@
-import type { Metadata } from "next"
-import Link from "next/link"
-import { EventsApiAccessGate } from "@/components/EventsApiAccessGate"
-import { CopyAgentBrief } from "./CopyAgentBrief"
-
-const API_BASE = process.env.NEXT_PUBLIC_API_URL ?? ""
-
-export const metadata: Metadata = {
-  title: "Events API",
-  description:
-    "Read events and RSVP on a user's behalf with a portal personal access token.",
-}
-
-type Method = "GET" | "POST" | "PATCH" | "DELETE"
-type Scope = "events:read" | "events:write" | "rsvp:write" | "venues:write"
-
-interface EndpointProps {
-  method: Method
-  path: string
-  scope: Scope
-  summary: string
-  params?: { name: string; type: string; note?: string }[]
-  body?: { name: string; type: string; note?: string }[]
-}
-
-const methodColor: Record<Method, string> = {
-  GET: "bg-emerald-100 text-emerald-700",
-  POST: "bg-sky-100 text-sky-700",
-  PATCH: "bg-amber-100 text-amber-700",
-  DELETE: "bg-rose-100 text-rose-700",
-}
-
-function Endpoint({
-  method,
-  path,
-  scope,
-  summary,
-  params,
-  body,
-}: EndpointProps) {
-  return (
-    <div className="border rounded-lg overflow-hidden bg-card">
-      <div className="flex items-center gap-3 px-4 py-3 border-b bg-muted/40">
-        <span
-          className={`text-xs font-mono font-semibold px-2 py-0.5 rounded ${methodColor[method]}`}
-        >
-          {method}
-        </span>
-        <code className="text-sm font-mono flex-1 truncate">{path}</code>
-        <span className="text-xs font-mono text-muted-foreground shrink-0">
-          {scope}
-        </span>
-      </div>
-      <div className="px-4 py-3 space-y-3">
-        <p className="text-sm text-muted-foreground">{summary}</p>
-        {params && params.length > 0 && (
-          <FieldList title="Query" items={params} />
-        )}
-        {body && body.length > 0 && <FieldList title="Body" items={body} />}
-      </div>
-    </div>
-  )
-}
-
-function FieldList({
-  title,
-  items,
-}: {
-  title: string
-  items: { name: string; type: string; note?: string }[]
-}) {
-  return (
-    <div>
-      <h4 className="text-xs font-semibold uppercase text-muted-foreground mb-1">
-        {title}
-      </h4>
-      <ul className="text-sm space-y-1">
-        {items.map((p) => (
-          <li key={p.name} className="font-mono text-xs">
-            <span className="text-foreground font-semibold">{p.name}</span>
-            <span className="text-muted-foreground"> · {p.type}</span>
-            {p.note && (
-              <span className="text-muted-foreground"> — {p.note}</span>
-            )}
-          </li>
-        ))}
-      </ul>
-    </div>
-  )
-}
-
-function Section({
-  title,
-  id,
-  action,
-  children,
-}: {
-  title: string
-  id?: string
-  action?: React.ReactNode
-  children: React.ReactNode
-}) {
-  return (
-    <section id={id} className="space-y-3">
-      <div className="flex items-center justify-between gap-3">
-        <h2 className="text-xl font-semibold">{title}</h2>
-        {action}
-      </div>
-      {children}
-    </section>
-  )
-}
-
-const ERRORS: { code: string; meaning: string }[] = [
-  { code: "401", meaning: "Missing, malformed, revoked, or expired key." },
-  { code: "403", meaning: "Key lacks the required scope for this route." },
-  {
-    code: "404",
-    meaning: "Resource is hidden from the caller or doesn't exist.",
-  },
-  {
-    code: "409",
-    meaning:
-      "Conflict — resource has dependent records (e.g. venue has events).",
-  },
-  { code: "422", meaning: "Body or query failed validation." },
-  { code: "429", meaning: "Rate limit exceeded — see Retry-After header." },
-]
-
-export default function ApiDocsPage() {
-  return (
-    <EventsApiAccessGate>
-      <ApiDocsBody />
-    </EventsApiAccessGate>
-  )
-}
-
-function ApiDocsBody() {
-  return (
-    <div className="flex-1 p-6 bg-background">
-      <div className="max-w-4xl mx-auto space-y-8">
-        <header className="space-y-2">
-          <h1 className="text-3xl font-semibold">Events API</h1>
-          <p className="text-muted-foreground">
-            Read events and RSVP on a user's behalf using a portal-issued
-            personal access token.
-          </p>
-        </header>
-
-        <Section
-          title="Quick start"
-          action={<CopyAgentBrief apiBase={API_BASE} />}
-        >
-          <div className="rounded-lg border bg-card divide-y text-sm">
-            <Row label="API base">
-              <code className="font-mono text-xs break-all">
-                {`${API_BASE}/api/v1`}
-              </code>
-              <span className="text-muted-foreground">
-                {" "}
-                — prefix only; every endpoint path lives beneath it
-              </span>
-            </Row>
-            <Row label="Auth header">
-              <code className="font-mono text-xs">
-                Authorization: Bearer eos_live_…
-              </code>
-            </Row>
-            <Row label="Get a token">
-              <Link
-                href="/portal/api-keys"
-                className="underline underline-offset-2"
-              >
-                /portal/api-keys
-              </Link>
-            </Row>
-            <Row label="Scopes available">
-              <code className="font-mono text-xs">events:read</code>
-              <span className="text-muted-foreground">, </span>
-              <code className="font-mono text-xs">events:write</code>
-              <span className="text-muted-foreground">, </span>
-              <code className="font-mono text-xs">rsvp:write</code>
-              <span className="text-muted-foreground">, </span>
-              <code className="font-mono text-xs">venues:write</code>
-            </Row>
-          </div>
-        </Section>
-
-        <Section title="Conventions">
-          <ul className="text-sm space-y-1 text-muted-foreground list-disc pl-5">
-            <li>
-              List endpoints return{" "}
-              <code className="font-mono text-xs">
-                {"{ results: T[], paging }"}
-              </code>
-              ; single-resource endpoints return the resource directly.
-            </li>
-            <li>Times are ISO-8601 with timezone. UUIDs are RFC-4122.</li>
-            <li>
-              Recurring events expand into virtual occurrences when{" "}
-              <code className="font-mono text-xs">start_after</code> is set.
-              Pass the occurrence's{" "}
-              <code className="font-mono text-xs">start_time</code> as{" "}
-              <code className="font-mono text-xs">occurrence_start</code> when
-              RSVPing to a single instance.
-            </li>
-          </ul>
-        </Section>
-
-        <Section title="Errors">
-          <div className="rounded-lg border bg-card divide-y">
-            {ERRORS.map((e) => (
-              <div
-                key={e.code}
-                className="flex items-baseline gap-3 px-4 py-2 text-sm"
-              >
-                <code className="font-mono text-xs w-10 shrink-0">
-                  {e.code}
-                </code>
-                <span className="text-muted-foreground">{e.meaning}</span>
-              </div>
-            ))}
-          </div>
-        </Section>
-
-        <Section title="Events" id="events">
-          <Endpoint
-            method="GET"
-            path="/api/v1/events/portal/events"
-            scope="events:read"
-            summary="List events visible to the caller."
-            params={[
-              { name: "popup_id", type: "uuid", note: "scope to one popup" },
-              { name: "kind", type: "string" },
-              { name: "venue_id", type: "uuid" },
-              { name: "track_id", type: "uuid" },
-              { name: "tags", type: "string[]" },
-              {
-                name: "start_after",
-                type: "ISO datetime",
-                note: "expands recurring masters into occurrences",
-              },
-              { name: "start_before", type: "ISO datetime" },
-              { name: "search", type: "string", note: "fuzzy title match" },
-              {
-                name: "rsvped_only",
-                type: "boolean",
-                note: "only events the caller has RSVPed to",
-              },
-              { name: "skip", type: "int" },
-              { name: "limit", type: "int", note: "max 100" },
-            ]}
-          />
-
-          <Endpoint
-            method="GET"
-            path="/api/v1/events/portal/events/{event_id}"
-            scope="events:read"
-            summary="Fetch one event. Includes the caller's RSVP status."
-            params={[
-              {
-                name: "occurrence_start",
-                type: "ISO datetime",
-                note: "scope RSVP lookup to one instance of a recurring event",
-              },
-            ]}
-          />
-
-          <Endpoint
-            method="PATCH"
-            path="/api/v1/events/portal/events/{event_id}"
-            scope="events:write"
-            summary="Update an event you own. Calendar-affecting changes (time, venue, title) bump the sequence and dispatch an iTIP UPDATE."
-            body={[
-              { name: "title", type: "string" },
-              { name: "content", type: "string" },
-              { name: "start_time", type: "ISO datetime" },
-              { name: "end_time", type: "ISO datetime" },
-              { name: "timezone", type: "string" },
-              {
-                name: "venue_id",
-                type: "uuid",
-                note: "clears custom_location_*",
-              },
-              {
-                name: "custom_location_name",
-                type: "string",
-                note: "clears venue_id",
-              },
-              { name: "custom_location_url", type: "string" },
-              { name: "cover_url", type: "string" },
-              { name: "meeting_url", type: "string" },
-              { name: "max_participant", type: "int" },
-              { name: "tags", type: "string[]" },
-              { name: "track_id", type: "uuid" },
-              {
-                name: "visibility",
-                type: "enum",
-                note: "public | private | unlisted",
-              },
-              { name: "status", type: "enum" },
-              { name: "host_display_name", type: "string" },
-            ]}
-          />
-
-          <Endpoint
-            method="POST"
-            path="/api/v1/events/portal/events/{event_id}/cancel"
-            scope="events:write"
-            summary="Soft-cancel an event you own. Sets status to CANCELLED and dispatches an iTIP CANCEL to attendees. There is no hard-delete."
-          />
-        </Section>
-
-        <Section title="Invitations" id="invitations">
-          <Endpoint
-            method="GET"
-            path="/api/v1/events/portal/events/{event_id}/invitations"
-            scope="events:read"
-            summary="List invitations for an event you own."
-          />
-
-          <Endpoint
-            method="POST"
-            path="/api/v1/events/portal/events/{event_id}/invitations"
-            scope="events:write"
-            summary="Bulk-invite humans by email. Emails must match existing humans in the tenant; unknown addresses come back under not_found. Owner-only."
-            body={[
-              {
-                name: "emails",
-                type: "string[]",
-                note: "1–1000 entries, case-insensitive",
-              },
-            ]}
-          />
-
-          <Endpoint
-            method="DELETE"
-            path="/api/v1/events/portal/events/{event_id}/invitations/{invitation_id}"
-            scope="events:write"
-            summary="Revoke a single invitation. Owner-only."
-          />
-        </Section>
-
-        <Section title="RSVP" id="rsvp">
-          <Endpoint
-            method="POST"
-            path="/api/v1/event-participants/portal/register/{event_id}"
-            scope="rsvp:write"
-            summary="RSVP to an event."
-            body={[
-              {
-                name: "occurrence_start",
-                type: "ISO datetime",
-                note: "required for recurring events, omit for one-offs",
-              },
-            ]}
-          />
-
-          <Endpoint
-            method="POST"
-            path="/api/v1/event-participants/portal/cancel-registration/{event_id}"
-            scope="rsvp:write"
-            summary="Cancel a previous RSVP."
-          />
-
-          <Endpoint
-            method="GET"
-            path="/api/v1/event-participants/portal/participants"
-            scope="events:read"
-            summary="List the caller's own RSVPs across events."
-          />
-        </Section>
-
-        <Section title="Venues" id="venues">
-          <Endpoint
-            method="GET"
-            path="/api/v1/event-venues/portal/venues"
-            scope="events:read"
-            summary="List active venues for a popup."
-            params={[
-              { name: "popup_id", type: "uuid", note: "required" },
-              {
-                name: "search",
-                type: "string",
-                note: "fuzzy title/location match",
-              },
-              { name: "skip", type: "int" },
-              { name: "limit", type: "int", note: "max 100" },
-            ]}
-          />
-
-          <Endpoint
-            method="POST"
-            path="/api/v1/event-venues/portal/venues"
-            scope="venues:write"
-            summary="Create a venue you own. Subject to the popup's humans_can_create_venues setting; may be created in PENDING status when the popup requires approval."
-            body={[
-              { name: "popup_id", type: "uuid" },
-              { name: "title", type: "string" },
-              { name: "description", type: "string" },
-              { name: "location", type: "string" },
-              { name: "formatted_address", type: "string" },
-              { name: "geo_lat", type: "float" },
-              { name: "geo_lng", type: "float" },
-              { name: "capacity", type: "int" },
-              { name: "image_url", type: "string" },
-              { name: "tags", type: "string[]" },
-              {
-                name: "booking_mode",
-                type: "enum",
-                note: "free | approval_required | unbookable",
-              },
-            ]}
-          />
-
-          <Endpoint
-            method="PATCH"
-            path="/api/v1/event-venues/portal/venues/{venue_id}"
-            scope="venues:write"
-            summary="Update a venue you own. The status field is ignored — re-approval lives in the backoffice."
-            body={[
-              { name: "title", type: "string" },
-              { name: "description", type: "string" },
-              { name: "location", type: "string" },
-              { name: "capacity", type: "int" },
-              { name: "image_url", type: "string" },
-              { name: "tags", type: "string[]" },
-              { name: "booking_mode", type: "enum" },
-            ]}
-          />
-
-          <Endpoint
-            method="DELETE"
-            path="/api/v1/event-venues/portal/venues/{venue_id}"
-            scope="venues:write"
-            summary="Delete a venue you own. Returns 409 if the venue still has non-cancelled events; remove or reassign them first."
-          />
-        </Section>
-      </div>
-    </div>
-  )
-}
-
-function Row({
-  label,
-  children,
-}: {
-  label: string
-  children: React.ReactNode
-}) {
-  return (
-    <div className="flex items-baseline gap-4 px-4 py-2">
-      <span className="text-xs font-semibold uppercase text-muted-foreground w-32 shrink-0">
-        {label}
-      </span>
-      <span className="text-sm">{children}</span>
-    </div>
-  )
-}
